KVM: arm64: selftests: Split arch_timer test code
mainline inclusion
from Linux 6.9-rc1
commit c20dd9e0695fc1e4da4ff01a111b0392862cfed6
category: feature
bugzilla: https://github.com/RVCK-Project/rvck/issues/184

--------------------------------

Split the arch-neutral test code out of aarch64/arch_timer.c
and put them into a common arch_timer.c. This is a preparation
to share timer test codes in riscv.

Suggested-by: Andrew Jones <ajones@ventanamicro.com>
